+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix a bug: unable to run scripts on systemts with legacy version of docker-compose
 - Fix a bug: unable to sign up if admin actived organization's additional fields with required
+- Fix a bug: undefined error in new member page
 
 ## v5.4.1 2022 May 23